ci: Build Postgres images preconfigured with logical replication for use in CI workflows (#3588)
This serves multiple goals:
- avoid repeating configuration code between CI workflows
- get rid of the need to run `psql` or `docker exec` commands in
addition to defining Postgres as a service in those CI workflows that
need it. See https://github.com/electric-sql/electric/pull/3600.
- https://github.com/electric-sql/electric/pull/3594 packages up
sync-service as a Docker image for use in CI. However, it cannot be
easily started as a service if the postgres service needs to be
configured afterwards

In other words, this together with building a Docker image for sync
service to use in CI workflows from the other PR helps avoid redoing
some of the build steps between different CI workflows.