Look for changes in any file under packages/sync-service
I've already missed the Dockerfile itself by trying to be overly
specific with the paths touched by a commit. Better be safe and take all
the entire contents of the package directory into account when deciding
on rebuilding the Docker image. After all, we have the Docker build
cache to avoid unnecessary rebuilds.