ci: Build a Docker image for sync-service to reuse in CI workflows (#3594)
Instead of duplicating caching and compilation logic for Electric in TS
tests, build a Docker image from sync-service's source code and run it
as a Docker container.

In a followup PR I'm going to also use the sync-service Docker container
when running Lux integration tests.

One positive outcome from this is that we no longer need a dummy
workflow to test that sync-service works when running in a Docker
container. It is now implicitly verified by running it in a Docker
container in existing workflows.

- [x] Replace the stub image tag derivation with a proper one that takes
into account all files relevant for the Docker image build of
sync-service