fix: ensure correct change handling and transformations on active move-ins (#3614)
- moves change processing from consumer to a separate module for easier
testing without stateful setup
- makes sure that changes only allowed to go through if they are
properly covering the in-flight snapshot
- makes sure that updates that don't touch a referenced value are still
converted from an update to an insert if it's the first time they are
seen.

On the second part, logic is as follows:
- when a change comes, if it doesn't interact with in-flight move-in
query results, then append using normal logic
- if a change references the value that was moved in (e.g. for
`parent_id IN (SELECT id FROM parent)` there have appeared parents with
ids `1` and `2`, and the change on the child table has `parent_id = 1`)
(only in-flight queries are considered), then:
- if the `pg_snapshot` is not known or is known and the txid is visible
in the snapshot, then we know that when query results return they will
have fresher view of the row and we need to keep the `NewRecord` from
the query results; so we skip this change on the stream
- we can apply this logic for when `pg_snapshot` is not known because
that means the query hasn't started resolving yet, and is going to be
definitely in the future from the current change.
- otherwise, if the change is not visible in `pg_snapshot`, then this
change has a fresher view of the row as compared to query results, and
so we append the change to the log (possibly converting updates to
inserts) and then, when applying query results, we'll skip the affected
key.

---------

Co-authored-by: rob <rob.acourt@gmail.com>
Co-authored-by: Rob A'Court <robacourt@users.noreply.github.com>