ci: Build Docker Hub images using GitHub Actions (#3617)
This reproduces our Buildkite pipeline on GitHub Actions and allows us
to retire Buildkite.

The most painful problem that this solves are the missing webhooks that
result in missed release images on Docker Hub. In a few words, when we
push a release commit to `main` that touches 4 or more components, and
so the commit itself gets 4+ tags, GitHub does not send a webhook to
Buildkite because of its limit on max tags. In such cases, we always had
to manually force-push only the `@core/sync-service:...` tag to trigger
a Buildkite pipeline.

The GitHub Actions workflow implemented here is verbose because it has
to orchestrate a single shared step, followed by a matrix that uses a
separate runner for building different platforms, followed by another
shared step that collects the digests from the matrix and combines them
into a single multi-arch tagged image on Docker Hub.

It may be possible to simplify it later on, since our next target is
migrating some or all of our runners to Blacksmith, and Blacksmith
promises to implement implicit forking of Docker builder into different
runners based on the passed platforms, kinda like Buildkite has been
doing.

Closes #3607.